+++ src/nmigate/customer/billing_record.py
@@ -1,6 +1,6 @@
 import uuid
 
-from nmigate import Nmi
+from ..nmi import Nmi
 
 
 class BillingRecord(Nmi):
