+++ src/nmigate/payment/__init__.py
@@ -1,0 +1,3 @@
+from .transactions import Transactions
+
+__all__ = ["Transactions"]
