+++ src/nmigate/subscription/__init__.py
@@ -1,0 +1,4 @@
+from .plans import Plans
+from .subscriptions import Subscriptions
+
+__all__ = ["Plans", "Subscriptions"]
